Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.c
@@ -16,10 +16,10 @@ int minEatingSpeed(int* piles, int pilesSize, int h) {
       
       int mid = left + (right - left) / 2;
       
-      long long needed_hours = 0;
+      int needed_hours = 0;
       
       for(int i=0; i < pilesSize; i++){
-        needed_hours = needed_hours + ((piles[i] + mid - 1) / mid);
+        needed_hours = needed_hours + ((long long)(piles[i] + mid - 1) / mid);
       }
       
       if(needed_hours <= h){
